Distributed Computations
I have implemented the ability to run a computation on multiple
machines. One machine can run as a server and others may connect as
clients. The clients receive work from the server, do the computations,
and send the results to the server. The server receives updates as the
clients progress on their assigned work.

ToDo List:
- Add the ability to connect to an already running server
- Add the ability to reconnect if something goes wrong (connection
error)
- Implement the merging of files after the entire computation is
complete
- Implement restarting or loading a previous computation
- The files completed by a client still need to be sent to the server
along with the corresponding data
-- Maybe the data should be stored in the file itself, this would be
helpful for loading
- Implement the usage of the files into a prediction algorithm to
determine whether a d8 or a d6 should be rolled to successfully cast any
given spell
- Show the likely hood to cast a set of spells based on current stats